fix photon leak in faker lib, fix examples
