+++ examples/simple-prisma/debug.ignore.ts
@@ -33,8 +33,6 @@ async function run() {
       },
     )
 
-    debugger
-
     console.log(data)
   } catch (err) {
     console.log(err)
